+++ contracts/test/foreign-proxy-optimism.test.ts
@@ -1,7 +1,15 @@
-const { ethers } = require("hardhat");
-const { time } = require("@nomicfoundation/hardhat-network-helpers");
-const { expect } = require("chai");
-const { toBigInt, ZeroAddress, zeroPadValue, toBeHex } = ethers;
+import { ethers } from "hardhat";
+import { time } from "@nomicfoundation/hardhat-network-helpers";
+import { expect } from "chai";
+import { toBigInt, ZeroAddress, zeroPadValue, toBeHex } from "ethers";
+import { HardhatEthersSigner } from "@nomicfoundation/hardhat-ethers/signers";
+import { 
+  AutoAppealableArbitrator, 
+  RealitioForeignProxyOptimism, 
+  MockRealitioHomeProxyOptimism, 
+  MockRealitio, 
+  MockCrossDomainMessenger
+} from "../typechain-types";
 
 const arbitratorExtraData = "0x85";
 const arbitrationCost = 1000;
@@ -15,7 +23,7 @@ const appealTimeOut = 180;
 const winnerMultiplier = 3000;
 const loserMultiplier = 7000;
 const loserAppealPeriodMultiplier = 5000;
-const gasPrice = toBigInt(80000000);
+const gasPrice = toBigInt(80000000n);
 const MAX_ANSWER = "115792089237316195423570985008687907853269984665640564039457584007913129639935";
 const maxPrevious = 2001;
 const messenger = "0x4200000000000000000000000000000000000007";
@@ -27,18 +35,18 @@ const oneETH = ethers.parseEther("1");
 
 const ZERO_HASH = zeroPadValue(toBeHex(0), 32);
 
-let arbitrator;
-let homeProxy;
-let foreignProxy;
-let realitio;
-let mockMessenger;
+let arbitrator: AutoAppealableArbitrator;
+let homeProxy: MockRealitioHomeProxyOptimism;
+let foreignProxy: RealitioForeignProxyOptimism;
+let realitio: MockRealitio;
+let mockMessenger: MockCrossDomainMessenger;
 
-let governor;
-let requester;
-let crowdfunder1;
-let crowdfunder2;
-let answerer;
-let other;
+let governor: HardhatEthersSigner;
+let requester: HardhatEthersSigner;
+let crowdfunder1: HardhatEthersSigner;
+let crowdfunder2: HardhatEthersSigner;
+let answerer: HardhatEthersSigner;
+let other: HardhatEthersSigner;
 
 describe("Cross-chain arbitration with appeals", () => {
   beforeEach("initialize the contract", async function () {
@@ -379,7 +387,7 @@ describe("Cross-chain arbitration with appeals", () => {
       value: appealCost,
     }); // This value doesn't fund fully.
     tx = await txFundAppeal;
-    txFee = (await tx.wait()).gasUsed * gasPrice;
+    txFee = (await tx.wait())!.gasUsed * gasPrice;
 
     newBalance = await getBalance(crowdfunder1);
     expect(newBalance).to.equal(
@@ -404,7 +412,7 @@ describe("Cross-chain arbitration with appeals", () => {
       .connect(crowdfunder1)
       .fundAppeal(arbitrationID, 533, { gasPrice: gasPrice, value: oneETH }); // Overpay to check that it's handled correctly.
     tx = await txFundAppeal;
-    txFee = (await tx.wait()).gasUsed * gasPrice;
+    txFee = (await tx.wait())!.gasUsed * gasPrice;
     newBalance = await getBalance(crowdfunder1);
     expect(newBalance).to.equal(
       oldBalance - toBigInt(3500) - txFee,
@@ -493,7 +501,7 @@ describe("Cross-chain arbitration with appeals", () => {
       gasPrice: gasPrice,
       value: arbitrationCost,
     });
-    const txFee = (await tx.wait()).gasUsed * gasPrice;
+    const txFee = (await tx.wait())!.gasUsed * gasPrice;
 
     const newBalance = await getBalance(requester);
     expect(newBalance).to.equal(
@@ -815,7 +823,7 @@ describe("Cross-chain arbitration with appeals", () => {
     expect(arbitration[3]).to.equal(50, "The answer should be 50");
   });
 
-  async function deployContracts(signer) {
+  async function deployContracts(signer: HardhatEthersSigner) {
     const Arbitrator = await ethers.getContractFactory("AutoAppealableArbitrator", signer);
 
     const arbitrator = await Arbitrator.deploy(String(arbitrationCost));
@@ -871,7 +879,7 @@ describe("Cross-chain arbitration with appeals", () => {
     };
   }
 
-  async function getBalance(account) {
+  async function getBalance(account: HardhatEthersSigner) {
     return account.provider.getBalance(await account.getAddress());
   }
 });
