+++ contracts/test/foreign-proxy-zksync.test.ts
@@ -1,7 +1,15 @@
-const { ethers } = require("hardhat");
-const { time } = require("@nomicfoundation/hardhat-network-helpers");
-const { expect } = require("chai");
-const { toBigInt, ZeroAddress, zeroPadValue, toBeHex } = ethers;
+import { ethers } from "hardhat";
+import { time } from "@nomicfoundation/hardhat-network-helpers";
+import { expect } from "chai";
+import { toBigInt, ZeroAddress, zeroPadValue, toBeHex } from "ethers";
+import { HardhatEthersSigner } from "@nomicfoundation/hardhat-ethers/signers";
+import { 
+  AutoAppealableArbitrator, 
+  MockRealitioHomeProxyZkSync, 
+  RealitioForeignProxyZkSync, 
+  MockRealitio, 
+  MockZkSync
+} from "../typechain-types";
 
 const arbitratorExtraData = "0x85";
 const arbitrationCost = 1000;
@@ -25,7 +33,7 @@ const appealTimeOut = 180;
 const winnerMultiplier = 3000;
 const loserMultiplier = 7000;
 const loserAppealPeriodMultiplier = 5000;
-const gasPrice = toBigInt(80000000);
+const gasPrice = toBigInt(80000000n);
 const MAX_ANSWER = "115792089237316195423570985008687907853269984665640564039457584007913129639935";
 const maxPrevious = 2001;
 
@@ -35,18 +43,18 @@ const foreignChainId = 5;
 const oneETH = ethers.parseEther("1");
 const ZERO_HASH = zeroPadValue(toBeHex(0), 32);
 
-let arbitrator;
-let homeProxy;
-let foreignProxy;
-let realitio;
-let mockZkSync;
+let arbitrator: AutoAppealableArbitrator;
+let homeProxy: MockRealitioHomeProxyZkSync;
+let foreignProxy: RealitioForeignProxyZkSync;
+let realitio: MockRealitio;
+let mockZkSync: MockZkSync;
 
-let governor;
-let requester;
-let crowdfunder1;
-let crowdfunder2;
-let answerer;
-let other;
+let governor: HardhatEthersSigner;
+let requester: HardhatEthersSigner;
+let crowdfunder1: HardhatEthersSigner;
+let crowdfunder2: HardhatEthersSigner;
+let answerer: HardhatEthersSigner;
+let other: HardhatEthersSigner;
 
 describe("Cross-chain arbitration with appeals", () => {
   beforeEach("initialize the contract", async function () {
@@ -150,7 +158,7 @@ describe("Cross-chain arbitration with appeals", () => {
     const tx = await foreignProxy
       .connect(requester)
       .requestArbitration(questionID, maxPrevious, { gasPrice: gasPrice, value: totalCost });
-    const txFee = (await tx.wait()).gasUsed * gasPrice;
+    const txFee = (await tx.wait())!.gasUsed * gasPrice;
 
     const newBalance = await getBalance(requester);
     expect(newBalance).to.equal(
@@ -463,7 +471,7 @@ describe("Cross-chain arbitration with appeals", () => {
     const tx = await foreignProxy
       .connect(other)
       .handleFailedDisputeCreation(questionID, await requester.getAddress(), { gasPrice: gasPrice, value: oneETH });
-    const txFee = (await tx.wait()).gasUsed * gasPrice;
+    const txFee = (await tx.wait())!.gasUsed * gasPrice;
 
     const newBalance = await await getBalance(other);
     expect(newBalance).to.equal(
@@ -547,7 +555,7 @@ describe("Cross-chain arbitration with appeals", () => {
     const tx = await foreignProxy
       .connect(other)
       .relayRule(questionID, await requester.getAddress(), { gasPrice: gasPrice, value: oneETH });
-    const txFee = (await tx.wait()).gasUsed * gasPrice;
+    const txFee = (await tx.wait())!.gasUsed * gasPrice;
 
     const newBalance = await getBalance(other);
     expect(newBalance).to.equal(
@@ -594,7 +602,7 @@ describe("Cross-chain arbitration with appeals", () => {
       .connect(crowdfunder1)
       .fundAppeal(arbitrationID, 533, { gasPrice: gasPrice, value: appealCost }); // This value doesn't fund fully.
     tx = await txFundAppeal;
-    txFee = (await tx.wait()).gasUsed * gasPrice;
+    txFee = (await tx.wait())!.gasUsed * gasPrice;
 
     newBalance = await getBalance(crowdfunder1);
     expect(newBalance).to.equal(
@@ -619,7 +627,7 @@ describe("Cross-chain arbitration with appeals", () => {
       .connect(crowdfunder1)
       .fundAppeal(arbitrationID, 533, { gasPrice: gasPrice, value: oneETH }); // Overpay to check that it's handled correctly.
     tx = await txFundAppeal;
-    txFee = (await tx.wait()).gasUsed * gasPrice;
+    txFee = (await tx.wait())!.gasUsed * gasPrice;
     newBalance = await getBalance(crowdfunder1);
     expect(newBalance).to.equal(
       oldBalance - toBigInt(3500) - txFee,
@@ -1054,11 +1062,11 @@ describe("Cross-chain arbitration with appeals", () => {
       .withArgs(arbitrator.target, arbitrationID, await other.getAddress(), "text");
   });
 
-  function getEmittedEvent(eventName, receipt, iface) {
-    return receipt.logs.map((log) => iface.parseLog(log)).find((parsed) => parsed.name === eventName);
+  function getEmittedEvent(eventName: any, receipt: any, iface: any) {
+    return receipt.logs.map((log: any) => iface.parseLog(log)).find((parsed: any) => parsed.name === eventName);
   }
 
-  async function deployContracts(signer) {
+  async function deployContracts(signer: HardhatEthersSigner) {
     const Arbitrator = await ethers.getContractFactory("AutoAppealableArbitrator", signer);
     const arbitrator = await Arbitrator.deploy(String(arbitrationCost));
 
@@ -1102,7 +1110,7 @@ describe("Cross-chain arbitration with appeals", () => {
     };
   }
 
-  async function getBalance(account) {
+  async function getBalance(account: HardhatEthersSigner) {
     return account.provider.getBalance(await account.getAddress());
   }
 });
